Reset global state when throwing b/c of circular refs
serialize() throws when it encounters a circular reference. However
because a global variable was being used, that global variable had
preserved state that would be reused in all future calls to serialize.

This caused a problem, particular with lists, which might be serialized
multiple times, being mutated in between serialization. In which case
they would always return the same (incorrect) value.

I fixed this by refactoring and making the global state a little more
explicit / easier to read, and providing a method `.end()` that does the
necessary clean up.

This will fix the failing canjs/canjs tests.